% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies_DEU.R
\docType{data}
\name{companies_DEU}
\alias{companies_DEU}
\title{100 largest companies of each state in Germany.}
\format{
A data frame with 1600 rows and 3 variables
}
\source{
\url{https://www.kaggle.com/peopledatalabssf/free-7-million-company-dataset}
}
\usage{
data("companies_DEU")
}
\description{
A dataset containing the 100 companies with the largest estimated number of employees
for each state (NUTS 1) in Germany.
Adapted from the 2019 Global Company Dataset made freely available by People Data Labs.
}
\keyword{datasets}
