% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies_ITA.R
\docType{data}
\name{companies_ITA}
\alias{companies_ITA}
\title{100 largest companies of each region in Italy}
\format{
A data frame with 2100 rows and 3 variables
}
\source{
\url{https://www.kaggle.com/peopledatalabssf/free-7-million-company-dataset}
}
\usage{
data("companies_ITA")
}
\description{
A dataset containing the 100 companies with the largest estimated number of employees
for each region (NUTS 2) in Italy.
Adapted from the 2019 Global Company Dataset made freely available by People Data Labs.
}
\keyword{datasets}
