% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_mask.R
\name{plot_region_mask}
\alias{plot_region_mask}
\title{plot_region_mask}
\usage{
plot_region_mask(df, bbox)
}
\arguments{
\item{df}{Dataframe for a given region obtained through [`gisco_get_nuts()`](https://rdrr.io/cran/giscoR/man/gisco_get_nuts.html), filtered by its NUTS id.}

\item{bbox}{The bounding box of the region.}
}
\value{
A ggplot object with the shape of the nuts region.
}
\description{
Plot the shape of the region to use as mask in `plot_region_wordcloud()`.
}
\examples{
\dontrun{
library(giscoR)
library(magrittr)
df <- giscoR::gisco_get_nuts(nuts_id = "DE2", resolution = "10")
bbox_region <- df \%>\% st_bbox()
plot_region_mask(df, bbox_region)
}

}
