% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_wordcloud.R
\name{plot_region_wordcloud}
\alias{plot_region_wordcloud}
\title{plot_region_wordcloud}
\usage{
plot_region_wordcloud(
  dataframe,
  mask,
  name_column_words,
  name_column_frequency,
  max_word_size = 4,
  rm_outside = TRUE
)
}
\arguments{
\item{dataframe}{The dataframe containing the input data for a single region.}

\item{mask}{The path to a png image of the region to be used as the shape of the wordcloud. Can be retreived by using `plot_region_mask()`.}

\item{name_column_words}{The name of the column in `dataframe` containing the words.}

\item{name_column_frequency}{The name of the column in `dataframe` containing the frequencies.}

\item{max_word_size}{The maximum size of the words in the wordcloud. At the minimum value `1` all the words are equally sized. Default is `4`.}

\item{rm_outside}{Whether to remove words that could not be fitted in the wordcloud area. If set to `FALSE`, these words will be stacked on top of each other at the centre of each region. Default is `TRUE`.}
}
\value{
A ggplot object with a randomly coloured wordcloud in the shape of the region ind `dataframe`.
}
\description{
Plot a wordcloud in the shape of a region.
}
\examples{
\dontrun{
companies_region <- data("companies_DEU") \%>\% filter(code == "DE2")
plot_region_wordcloud(companies_region, "./mask.png", companies_region[["name"]], companies_region[["employees"]])
}

}
